base/common/src/test/java/com/netscape/certsrv/system/AdminSetupResponseTest.java
package com.netscape.certsrv.system;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

public class AdminSetupResponseTest {

    private static AdminSetupResponse before = new AdminSetupResponse();

    @Before
    public void setUpBefore() {
        before.setAdminCert(new SystemCertData());
    }

    @Test
    public void testXML() throws Exception {
        // Act
        String xml = before.toXML();
        System.out.println("XML (before): " + xml);

        AdminSetupResponse afterXML = AdminSetupResponse.fromXML(xml);
        System.out.println("XML (after): " + afterXML.toXML());

        // Assert
        Assert.assertEquals(before, afterXML);
    }

    @Test
    public void testJSON() throws Exception {
        // Act
        String json = before.toJSON();
        System.out.println("JSON (before): " + json);

        AdminSetupResponse afterJSON = AdminSetupResponse.fromJSON(json);
        System.out.println("JSON (after): " + afterJSON.toJSON());

        // Assert
        Assert.assertEquals(before, afterJSON);
    }

}
